+++ hata/discord/application/application/tests/test__parse_redirect_urls.py
@@ -4,13 +4,72 @@ from ..fields import parse_redirect_urls
 
 
 def _iter_options():
-    yield {}, None
-    yield {'redirect_uris': None}, None
-    yield {'redirect_uris': []}, None
-    yield {'redirect_uris': [None]}, None
-    yield {'redirect_uris': ['a']}, ('a', )
-    yield {'redirect_uris': ['a', None]}, ('a', )
-    yield {'redirect_uris': ['b', 'a']}, ('a', 'b')
+    yield (
+        {},
+        None,
+    )
+    
+    yield (
+        {
+            'redirect_uris': None,
+        },
+        None,
+    )
+    
+    yield (
+        {
+            'redirect_uris': [],
+        },
+        None,
+    )
+    
+    yield (
+        {
+            'redirect_uris': [
+                None,
+            ],
+        },
+        (
+            None,
+        ),
+    )
+    
+    yield (
+        {
+            'redirect_uris': [
+                'a',
+            ],
+        },
+        (
+            'a',
+        ),
+    )
+    
+    yield (
+        {
+            'redirect_uris': [
+                'a',
+                None,
+            ],
+        },
+        (
+            None,
+            'a',
+        ),
+    )
+    
+    yield (
+        {
+            'redirect_uris': [
+                'b',
+                'a',
+            ],
+        },
+        (
+            'a',
+            'b',
+        ),
+    )
 
 
 @vampytest._(vampytest.call_from(_iter_options()).returning_last())
@@ -25,6 +84,13 @@ def test__parse_redirect_urls(input_data):
     
     Returns
     -------
-    output : `None | tuple<str>`
+    output : `None | tuple<None | str>`
     """
-    return parse_redirect_urls(input_data)
+    output = parse_redirect_urls(input_data)
+    vampytest.assert_instance(output, tuple, nullable = True)
+    
+    if (output is not None):
+        for element in output:
+            vampytest.assert_instance(element, str, nullable = True)
+    
+    return output