+++ hata/discord/application/application/tests/test__put_redirect_urls.py
@@ -4,10 +4,47 @@ from ..fields import put_redirect_urls
 
 
 def _iter_options():
-    yield None, False, {}
-    yield None, True, {'redirect_uris': []}
-    yield ('a', ), False, {'redirect_uris': ['a']}
-    yield ('a', ), True, {'redirect_uris': ['a']}
+    yield (
+        None,
+        False,
+        {},
+    )
+    
+    yield (
+        None,
+        True,
+        {
+            'redirect_uris': [],
+        },
+    )
+    
+    yield (
+        (
+            None,
+            'a',
+        ),
+        False,
+        {
+            'redirect_uris': [
+                None,
+                'a',
+            ],
+        },
+    )
+    
+    yield (
+        (
+            None,
+            'a',
+        ),
+        True,
+        {
+            'redirect_uris': [
+                None,
+                'a',
+            ],
+        },
+    )
 
 
 @vampytest._(vampytest.call_from(_iter_options()).returning_last())
@@ -17,8 +54,9 @@ def test__put_redirect_urls(input_value, defaults):
     
     Parameters
     ----------
-    input_value : `None | tuple<str>`
+    input_value : `None | tuple<None | str>`
         Value to serialize.
+    
     defaults : `bool`
         Whether fields with the default values should be included as well.