+++ hata/discord/application/application/tests/test__validate_redirect_urls.py
@@ -7,11 +7,68 @@ def _iter_options__passing():
     url_0 = 'https://orindance.party/'
     url_1 = 'https://www.astil.dev/project/hata/'
     
-    yield None, None
-    yield [], None
-    yield url_0, (url_0, )
-    yield [url_0], (url_0, )
-    yield [url_1, url_0], (url_0, url_1)
+    yield (
+        None,
+        None,
+    )
+    
+    yield (
+        [],
+        None,
+    )
+    
+    yield (
+        url_0,
+        (
+            url_0,
+        ),
+    )
+    
+    yield (
+        None,
+        None,
+    )
+    
+    yield (
+        [
+            None,
+        ],
+        (
+            None,
+        ),
+    )
+    
+    
+    yield (
+        [
+            url_0,
+        ],
+        (
+            url_0,
+        ),
+    )
+    
+    yield (
+        [
+            None,
+            url_0,
+        ],
+        (
+            None,
+            url_0,
+        ),
+    )
+    
+    yield (
+        [
+            url_1,
+            url_0,
+        ],
+        (
+            url_0,
+            url_1,
+        ),
+    )
 
 
 def _iter_options__type_error():
@@ -32,10 +89,17 @@ def test__validate_redirect_urls(input_value):
     
     Returns
     -------
-    output : `None | tuple<str>`
+    output : `None | tuple<None | str>`
     
     Raises
     ------
     TypeError
     """
-    return validate_redirect_urls(input_value)
+    output = validate_redirect_urls(input_value)
+    vampytest.assert_instance(output, tuple, nullable = True)
+    
+    if (output is not None):
+        for element in output:
+            vampytest.assert_instance(element, str, nullable = True)
+    
+    return output